Commit for Record
Inputting images for the program is a bit of a tedius process which requires you to edit a couple variables in the code according to the number and extension of the images you have.

This will be fixed in a future update.

I have added a few images to show the result.